Organize modules (#138)
* scaffolding for feature module and shared module

This is where the fun begins - Anakin Skywalker

* still not working, but some structure is completed

We're just components sir, we're meant to be overused.

* we're making progress

* It loads!

still has some errors, but I'm going in the right direction.

Perhaps the archives are incomplete

* finished home module (for now); creating feature modules per page

Dew it

* fixed some bugs

not to worry, we're still managing half a website

* still broken, but things are a bit more organized

all part of the plan... THE plan.

* it compiles! (with runtime errors)

I need to fix runtime errors, remove unnecessary imports, finish moving all other components, and add lazy loading

* new modules, and some import cleaning

I've got a bad feeling about this

* organization part 1 is complete

General <strong> tag, you are a bold one!

* moved modules. began to lazy load

* organization is done, basic lazy loading complete

you were supposed to lazy load the modules! Not leave them together!

* a little more lazy loading

good bye old branch, and may the force be with you.